Moved functions to separate file
